adding step wise financials functions
